Create a function for just opening the associated spec file
Also, if rg is used to find the specs, consider it an error if several
spec files are found. This to reduce the complexity of handling multiple
specs in different functions that could receive that. The
Open_Spec_File() function for example wouldn't know what to do with
those. It could open all of them in a loop, but what if there are 100?
Handling all these cases is complicated.